requirements.txt
Flask==2.3.3
python-dotenv==1.0.1
langchain==0.2.7
langchain-google-genai==0.1.0
google-api-core==2.12.0
google-genai==0.1.1
tenacity==8.2.2
requests==2.31.0
Jinja2==3.1.2
